Add WitnessFilter
