+++ modules/core/src/main/scala-2/dev/profunktor/redis4cats/TypeInqualityCompat.scala
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2018-2021 ProfunKtor
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package dev.profunktor.redis4cats
+
+trait TypeInequalityCompat {
+
+  /**
+    * Type inequality
+    *
+    * Credits: https://stackoverflow.com/a/6929051
+    */
+  sealed class =!=[A, B]
+
+  object =!= extends NEqualLowPriority {
+    implicit def nequal[A, B]: =!=[A, B] = new =!=[A, B]
+  }
+
+  trait NEqualLowPriority {
+    implicit def equal[A]: =!=[A, A] = sys.error("should not be called")
+  }
+}
