+++ modules/effects/src/main/scala/dev/profunktor/redis4cats/pipeline.scala
@@ -45,10 +45,9 @@ object pipeline {
       * Same as @exec, except it filters out values of type Unit
       * from its result.
       */
-    def filterExec[T <: HList, R <: HList, S <: HList](commands: T)(
-        implicit w: Witness.Aux[T, R],
-        f: Filter.Aux[R, S]
-    ): F[S] = Runner[F].filterExec(ops)(commands)
+    def filterExec[T <: HList](commands: T)(
+        implicit w: WitnessFilter[T]
+    ): F[w.S] = Runner[F].filterExec(ops)(commands)
 
     /***
       * Exclusively run Redis commands as part of a pipeline (autoflush: disabled).
